Replace media/manga item cell with no-op placeholder
