+++ CMakeLists.txt
@@ -120,7 +120,6 @@ set(APP_SOURCES
     src/view/search_tab.cpp
     src/view/settings_tab.cpp
     src/view/media_detail_view.cpp
-    src/view/media_item_cell.cpp
     src/view/recycling_grid.cpp
     src/view/rotatable_box.cpp
     src/view/rotatable_image.cpp
